修复bug

+++ NestSound/Src/NSBase/Controller/NSPlayMusicViewController.h
@@ -24,7 +24,7 @@
 @property (nonatomic,assign) int geDanID;
 @property (nonatomic,strong) NSMutableArray * songAry;
 @property (nonatomic,assign) NSInteger songID;
-@property (nonatomic,assign) int isShow;
+@property (nonatomic,assign) long isShow;
 - (void)playMusicUrl:(NSString *)musicUrl;
 
 @end
